color float
Former-commit-id: 643809843d98d28fbfb025d6a0c37a31833c1d22 [formerly f1dd0d9f17828ac999db5b13dd406e8d14110323] [formerly 932bceeca8f5c078ee72978dd55096b0abcae514 [formerly 7025739ebe56e553bea7d95049d6802996a84e83]] [formerly 9ab9ed56b23297511337ec8399b4a660fdfc1c7e [formerly bef083a5718a14426050346dba20a67822256f7d] [formerly 02b7b3b864c394dff9e84ef3b45b8e99f16c3163 [formerly aa3affb7966f263aefefd31b9602e006d5a4cf53]]] [formerly aef71d5255b782f23951bca3e3bd44b307451070 [formerly 65f600207d9e05254db5ec7f692d55542fcb60af] [formerly 8e91b0d70d3bd05521017e98c07fa34dc2bd9907 [formerly eeaa50a50ca22ef40df4c8f0a3a33ee4b357c1e9]] [formerly 80454b9035e07c77085851e584cc4472b3b8e783 [formerly cefe05f6ad14dead0f44fd12f76f4a408495e6a6] [formerly bdf79b3be6a961b25a052899d2cce17c9fdf35dd [formerly ce19ba3e8d83208175ebe4f4b2f8d93b5e3c3f3c]]]] [formerly 2a676987f28817b1228b25ae78d506adc9f8348f [formerly ae9aaaaf5be0cb300bf2a8219d98719ac454131f] [formerly fd1233132c456100e05c8042e9761f4d2d0978d2 [formerly f54befb093c8694bac2eb681c88cf9b06b9d13d0]] [formerly 8e01df9aaa0142009a9d1f6816c4047d265e47ad [formerly b040f9819172d33354efdde35a624cbb9d485f63] [formerly 1aa7340a81bbaf10e732cdcb3cf9762e98f3d676 [formerly e605fae31548af4dcde7a37b0db82f7ac20dcfc7]]] [formerly 3668271489bb14122b8ef8277ce07b82456b5019 [formerly 4ac1d6013038b8979d63a4b5a09c299e4dd292b4] [formerly 1e6532db36728425394279ad484ea5b2b9d644a3 [formerly cd5bdbc3327a81a810d8f436beb8bb743b93664f]] [formerly 30665a110d6e1b088e6c7cc116336f22616d848d [formerly 365a5b30953152ba3dfd44c72b874c956202e9b0] [formerly 6148f69fdb6cfd2a78bf52c48ff0615e8f0aa020 [formerly c05ef89e5c3175d7f95b0a915201fe4abd3d8ad6]]]]]
Former-commit-id: 06f9f7a20953328351b9cf2f77a93482bf9fb71d [formerly 7674a91c95b8265241ee4bb7804f386ac76e1a21] [formerly 42defca61f310613e3c417fef43cc694a5274818 [formerly 3bb4d66761cc7727cc8f8e74a99c61ffa4aaad15]] [formerly 060eb9dc649a2e58dbe8fd0e33e6bd3ec2f1df4e [formerly 88759cd00b5c2a9b00cd847237bea7dd858e6e9a] [formerly 8fb0b6a8d7251d165e662694f3708e49f77cbf3b [formerly 592d6a245ac58ea8062adba39b5deba9a8c1a964]]] [formerly 6a7bf0031cf3911b01a4dea6933331f91a48777d [formerly 3e28bc9442c50a2f2e706fffbfa4de19cc8c16a8] [formerly d7fd2a4f3538443477c3956c88b6958c5c2d6989 [formerly e7f081da09709c8ab3455ed40e6f6895be8d2132]] [formerly 44f0de9bb6e958563101f73eca4ec455e1b3a6da [formerly cb7d7ce96a6d68ade047cb3457dccb59d0ba3b4d] [formerly 4c69e0fd57b4c5862353649a2fffc5495731d41b [formerly 35bd72c9b6ea06d9192cd62696fec0786000c2a7]]]]
Former-commit-id: fcd1a54a581f060ea346c2962f8d1a32874622c8 [formerly f9211591f9950315f62b91f5024820de9f28047c] [formerly ca075d110e81eeea9f983271d98c63d5ca654096 [formerly 883c022764e21e6e15feb55256da9e3737c9369f]] [formerly 35e23792b5973af110d6037e4c41238a49955ce5 [formerly 7b3ea878a0521edc3809ed20e103e7ecbc994e36] [formerly 36a1b1c6ec35ae3c7417e7558aeb3eda6cc68b36 [formerly ed85455e58f095307fa55d22cf4eeca3db7866d6]]]
Former-commit-id: 402ec34f89dbdabdea1093856a587599661fc1c3 [formerly 12b815850187b058c0901c32ea15ef48c2ecbfc5] [formerly f4cd5e55b99b9a9a4f41f37947f6769cc80b2123 [formerly 898dfc2ec2f42eed1c6b0af1fa6cdf6e56357ffa]]
Former-commit-id: 1cbc10aa016bedb63bc9d9f695bb4af72a6daa68 [formerly dcb6cce89307d9cb291299ff48b3cd2b16d54122]
Former-commit-id: b252102e7c4f2151f0414e44e9f37bbf820b93a9

+++ src/color.h
@@ -6,13 +6,13 @@ namespace sren {
 
 class Color {
  public:
-  int r{};
-  int g{};
-  int b{};
-  int a{};
+  float r{};
+  float g{};
+  float b{};
+  float a{};
   Color() = default;
-  Color(int _r, int _g, int _b) : r{_r}, g{_g}, b{_b}, a{255} {}
-  Color(int _r, int _g, int _b, int _a) : r{_r}, g{_g}, b{_b}, a{_a} {}
+  Color(float _r, float _g, float _b) : r{_r}, g{_g}, b{_b}, a{255} {}
+  Color(float _r, float _g, float _b, float _a) : r{_r}, g{_g}, b{_b}, a{_a} {}
 };
 
 }  // namespace sren
